address comments and remove hn from ci
